requirements.txt
scikit-learn>=0.18.1
numpy
minepy>=1.2.0
pandas>=0.20.1
flask>=0.11.1
werkzeug>=0.11.10
flask-jwt-extended>=2.1.1
mongoengine>=0.13.0
pymongo>=3.2.2
flask-cors>=3.0.2
pandas>=0.20.1
scipy
flask_socketio>=2.9.0
eventlet
keras>=2.0.9
sklearn
redis
simplejson
hyperas
hyperopt
pillow
h5py>=2.7.0
kubernetes
scikit-image
port_for
kubernetes
grpcio
synonyms
requests
pyyaml
kubernetes
port_for
docker
Pillow
jupyterhub==0.8.1
authenticator
dockerspawner
gitpython
cookiecutter
psutil
htpasswd
pyjwt
Cython
ed25519
rocketchat_API
pycrypto
celery
anytree
ipfsapi
jupyterhub-kubespawner
tensorflow==1.11.0
flask-restplus
qiniu
elasticsearch
schedule
jieba
pypinyin
Augmentor
kaggle
bs4
elasticsearch5
PyMySQL
selenium
zeep
xmltodict
bson

